Model support: Separated the pose of a model from the definition of the model, encapsulating the pose in the newly introduced class `AnimPoseT`.
This is a preparatory step for completing the support of animation channels and animation blending, where a pose will be composed of possibly several animation sequences playing on different channels, and animation sequences that are blended into each other.

This change also improves the existing code and design, as separate concepts are now cleanly separated in code.
It also improves caching, because all cached data is encapsulated in `AnimPoseT` where it is up to the user code to share or reuse the data across models and/or video frames.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@400 5fe0b3f3-48df-4727-b460-e6dda9226989