Model support: Removed the obsolete class `ModelMdlT`. In its place, HL1 mdl models are now loaded via the `LoaderHL1mdlT` into the newer and better class `CafuModelT`.
Also fixed the orientation of stencil shadow polygons (silhouette and "endcaps") of face occluders in the world BSP tree, making their orientation consistent to those used in `CafuModelT`. Without this, the sign of the stencil count of the world shadows mismatches that of the models shadows, yielding artifacts whenever model shadows and world shadows overlap.

References #78.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@404 5fe0b3f3-48df-4727-b460-e6dda9226989