Model code: Removed support for "gts" (given tangent space) from the .ase loader.
Meshes loaded with the .ase loader are set to use the `HARD` tangent-space method by default, as this helps with many older and/or simpler models that are not immediately edited (converted to `cmdl`) in the Model Editor.

"Given tangent space" is when our model code doesn't compute the normals and tangent vectors itself, but relies on those found in the file.
This, however, is problematic in many regards (requires extra storage, works only for non-animated models, can be incomplete (comes with normals, but not with tangents), etc.), and thus support for "gts" is removed from our model code.

References #100.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@481 5fe0b3f3-48df-4727-b460-e6dda9226989