Net code: Fix direct public access to `EntityStateT` instances.
`EntityStateT`s are private members of entities that we strive to entirely remove in the course of #113 as described at http://www.cafu.de/forum/viewtopic.php?p=4767#p4767.
References to them used to be passed around publicly in user code, violating the encapsulation of entity internals and effectively preventing further refactoring of the code.
The changes in this revision improve the class interfaces to enhance encapsulation and to make further refactoring possible.

References #113.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@530 5fe0b3f3-48df-4727-b460-e6dda9226989