Engine core, net code: Fixed a nasty bug in `EngineEntityT::ParseServerDeltaUpdateMessage()`.
We cannot rely on `GetState()` to produce the state as it was sent from the server in an earlier entity update message, because the state of a client entity is continuously updated (e.g. in `PostDraw()`), e.g. with simple extrapolations.
Have to properly store newly updated states in the `m_OldStates` array instead, and use those for delta-decompression.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@606 5fe0b3f3-48df-4727-b460-e6dda9226989