Game code: Switched from extrapolation back to interpolation.
Unfortunately, extrapolation is less useful and worthwhile than it initially seems, because when it is applied to entity origins, all kinds of unwanted side effects can happen:
  - items that fall to the floor penetrate the ground plane, then resurface;
  - the wings of closing doors bump into each other, then retract;
  - lifts overshoot their halting position;
  - grenades that are deflected by walls briefly vanish into them;
  - etc.

All this is fixed by switching (back) to interpolation, at the cost of the visual (but barely noticeable) delay of one server update.

This revision also largely renames "extrapolation" to "interpolation", which is now more prominent again and anyway a more familiar word.
Unfortunately, except for the rather vague and broad "approximation", I've not found a well fitting generic term that covers both.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@624 5fe0b3f3-48df-4727-b460-e6dda9226989